pyproject.toml
[project]
name = "quantum-algorithm-explorer"
version = "0.1.0"
description = "AI-powered quantum algorithm explorer using Claude agents and Qiskit simulation"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "qiskit>=1.0",
    "qiskit-aer>=0.17.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
quantum-explorer = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
